refactor(cli): merge fuzzer subcommand
