+++ crates/cli/fuzzer/README.md
@@ -1,4 +1,4 @@
-# evm2-fuzzer
+# evm2 fuzzer
 
 Differential fuzzer for evm2 against revm.
 
@@ -9,13 +9,13 @@ state changes.
 Run generated cases:
 
 ```sh
-cargo run -q -p evm2-fuzzer -- --seed 1 --cases 100000 -j 0
+cargo run -q -p evm2-cli -- fuzzer --seed 1 --cases 100000 -j 0
 ```
 
 Run by duration with a random seed:
 
 ```sh
-cargo run -q -p evm2-fuzzer -- --duration 5m -j 0
+cargo run -q -p evm2-cli -- fuzzer --duration 5m -j 0
 ```
 
 `-j`/`--threads` controls worker threads. `0` uses logical cores. The fuzzer
@@ -24,7 +24,7 @@ prints the seed for replayability.
 Force an evm2 dispatch backend with `EVM2_DISPATCH_BACKEND`:
 
 ```sh
-EVM2_DISPATCH_BACKEND=packed cargo run -q -p evm2-fuzzer -- --duration 1m -j 0
+EVM2_DISPATCH_BACKEND=packed cargo run -q -p evm2-cli -- fuzzer --duration 1m -j 0
 ```
 
 Accepted backends are `auto`, `tco`, `packed`, `single_return`, and `unpacked`.
@@ -34,25 +34,25 @@ Accepted backends are `auto`, `tco`, `packed`, `single_return`, and `unpacked`.
 Failing generated cases are written to:
 
 ```text
-crates/fuzzer/corpus/failures/
+crates/cli/fuzzer/corpus/failures/
 ```
 
 Replay one saved case:
 
 ```sh
-cargo run -q -p evm2-fuzzer -- replay crates/fuzzer/corpus/failures/case-....json
+cargo run -q -p evm2-cli -- fuzzer replay crates/cli/fuzzer/corpus/failures/case-....json
 ```
 
 Replay every JSON case in a corpus directory:
 
 ```sh
-cargo run -q -p evm2-fuzzer -- corpus crates/fuzzer/corpus/failures
+cargo run -q -p evm2-cli -- fuzzer corpus crates/cli/fuzzer/corpus/failures
 ```
 
 Minimize a reproducing case:
 
 ```sh
-cargo run -q -p evm2-fuzzer -- minimize crates/fuzzer/corpus/failures/case-....json
+cargo run -q -p evm2-cli -- fuzzer minimize crates/cli/fuzzer/corpus/failures/case-....json
 ```
 
 Import the `evm-protobuf-fuzzer` corpus as replayable JSON cases:
@@ -61,7 +61,7 @@ Import the `evm-protobuf-fuzzer` corpus as replayable JSON cases:
 ./scripts/import_evm_protobuf_corpus.py \
   /path/to/corp-evm-protobuf-fuzzer.tar.xz \
   --output tmp/evm-protobuf-import --clean
-cargo run -q -p evm2-fuzzer -- corpus tmp/evm-protobuf-import
+cargo run -q -p evm2-cli -- fuzzer corpus tmp/evm-protobuf-import
 ```
 
 The import is intentionally lossy where the fuzzer case format has no field for
@@ -71,7 +71,7 @@ cases are tagged in `features`.
 ## Coverage report
 
 Generate an HTML coverage report from fuzzer execution, excluding the fuzzer
-crate itself:
+module itself:
 
 ```sh
 ./scripts/fuzzer_coverage.py --duration 3m -j 0 --open
