+++ crates/cli/src/fuzzer/cli.rs
