+++ crates/cli/src/fuzzer/minimize.rs
@@ -1,4 +1,4 @@
-use crate::{
+use crate::fuzzer::{
     backend::EvmBackend,
     case::{EvmCase, TARGET},
 };
