+++ crates/cli/src/fuzzer/rng.rs
