--- crates/fuzzer/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "evm2-fuzzer"
-description = "Differential fuzzing harness for evm2"
-
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-categories.workspace = true
-keywords.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-evm2.workspace = true
-alloy-consensus = { workspace = true, features = ["secp256k1"] }
-alloy-eips = { workspace = true, features = ["serde", "std"] }
-alloy-primitives = { workspace = true, features = ["secp256k1", "serde", "std"] }
-clap = { version = "4.6", features = ["derive"] }
-rand.workspace = true
-secp256k1 = { workspace = true, features = ["global-context", "recovery"] }
-revm = { version = "38", default-features = true }
-serde = { workspace = true, features = ["derive", "std"] }
-serde_json = { workspace = true, features = ["std"] }
